.github/dependabot.yml: Enable dependabot (#169)
Enables dependabot in this repo.

This GitHub action will automatically create pull requests and
summarize the dependency details. Because it is a pull request,
the CI system will validate the dependency update in the pull
request.

Configures dependabot for:

1. PIP module updates
2. Submodule updates
3. GitHub action updates

- [ ] Breaking change?
- Will this change break pre-existing builds or functionality without
action being taken?
  **No**

Verified config in other repos.

N/A - Impacts this repo

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>