Add global CodeQL filter file to repo
The CodeQL Analyze build plugin allows a global configuration file
to be specified for a repo that can define file patterns against
CodeQL rules to include and exclude results for the given file and
rule.

More details are in the plugin readme:
https://github.com/microsoft/mu_basecore/blob/release/202208/.pytool/Plugin/CodeQL/Readme.md#global-configuration

This change calls out addition of this file for future usage.

The global configuration file is useful in mu_basecore (vs specifying
the filters in individual package CI YAML files) because the repo
has high reuse. Integration repos can easily see all of the filter
activity in one file that is easy to inspect and include in their
build if needed.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>