python: resolve parser deprecation warnings (#340)
base_parser.py:BaseParser - `SetBaseAbsPath()` and `SetPackagePaths()`
have been deprecated in favor of `SetEdk2Path()`.

This deprecation is to improve performance when parsing files.
Originally the class instantiated it's own Edk2Path object, which is an
expensive operation as it parses the workspace to confirm Packages Path
and Package validity. While this is not noticed for a single call,
typically the parsers are used to parse multiple files at once.

This change allows (and eventually enforces) that the Edk2Path object be
passed to the parser rather than instantiated each time, so that only
one Edk2Path object is created for `N` number of file parses.